pyproject.toml
[project]
name = "lite-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.74.8",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
